pyproject.toml
[project]
name = "evaluation-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
	"datalab-python-sdk>=0.5.0",
	"fastapi>=0.135.3",
	"httpx>=0.28.1",
	"openai>=2.31.0",
	"python-dotenv>=1.2.2",
	"python-multipart>=0.0.24",
	"requests>=2.33.1",
	"sqlalchemy>=2.0.41",
	"asyncpg>=0.30.0",
	"uvicorn>=0.44.0",
]
